Refactor: move logo components to components folder #344
